+++ src/GafferArnoldModule/GafferArnoldModule.cpp
@@ -46,6 +46,7 @@
 #include "GafferArnold/ArnoldImager.h"
 #include "GafferArnold/ArnoldLight.h"
 #include "GafferArnold/ArnoldMeshLight.h"
+#include "GafferArnold/ArnoldOperator.h"
 #include "GafferArnold/ArnoldOptions.h"
 #include "GafferArnold/ArnoldRender.h"
 #include "GafferArnold/ArnoldShader.h"
@@ -134,6 +135,15 @@ BOOST_PYTHON_MODULE( _GafferArnold )
 			.value( "Replace", ArnoldImager::Mode::Replace )
 			.value( "InsertFirst", ArnoldImager::Mode::InsertFirst )
 			.value( "InsertLast", ArnoldImager::Mode::InsertLast )
+		;
+	}
+
+	{
+		scope s = GafferBindings::DependencyNodeClass<ArnoldOperator>();
+		enum_<ArnoldOperator::Mode>( "Mode" )
+			.value( "Replace", ArnoldOperator::Mode::Replace )
+			.value( "InsertFirst", ArnoldOperator::Mode::InsertFirst )
+			.value( "InsertLast", ArnoldOperator::Mode::InsertLast )
 		;
 	}
 }